project/rdocai/requirements.txt
pillow
tqdm
numpy
datasets
transformers
torch
bitsandbytes